bugfix: Remove jrtClassPathCAches to try fix -release issues (#7697)
I can't find the actual problem where it happens unfortunately, so this is my best guess on what can help.
This might anyway be needed for older Scala 2.13 versions, but we could try to fix it in later.

I tested this with the debugger and clearing the cache actually fixed things.

(cherry picked from commit 8a7d36650029fbf69bc22eec8641e35cf3799948)