improvement: Do not return the full formatting (#7760)
If we return the full string, this gets added to LLM context and it might not even apply the correct formatting.

This might not work for the standalone client, but there is no way of knowing if we have a "fake" client.

(cherry picked from commit 6cb0730d06c2e95b203d5226d3a374c92a7e4b8f)